--- packages/plugin-bootstrap/__tests__/readme_tests.md
@@ -1,83 +1,0 @@
-# Bootstrap Plugin Test Suite
-
-## Overview
-
-This directory contains a comprehensive suite of tests for the Eliza Bootstrap Plugin. The test suite covers all major components of the plugin including actions, providers, evaluators, services, and event handling logic.
-
-## Test Coverage Status
-
-| Test File            | Components Tested                   | Coverage | Status                                  |
-| -------------------- | ----------------------------------- | -------- | --------------------------------------- |
-| `actions.test.ts`    | All actions in the bootstrap plugin | High     | ✅ All 30 tests pass                    |
-| `evaluators.test.ts` | Reflection evaluator                | High     | ✅ All 7 tests pass, with some warnings |
-| `logic.test.ts`      | Event handlers, message processing  | High     | ⚠️ 14/15 tests pass                     |
-| `plugin.test.ts`     | Plugin structure, initialization    | Medium   | ⚠️ 7/14 tests pass                      |
-| `providers.test.ts`  | Providers for state & data          | Low      | ⚠️ 4/20 tests pass                      |
-| `services.test.ts`   | Task service, scenario service      | Low      | ⚠️ 9/23 tests pass                      |
-
-## Strengths
-
-1. **Comprehensive Actions Testing**: All action tests are passing with complete coverage of edge cases, validation logic, and action handling.
-
-2. **Strong Evaluator Tests**: The reflection evaluator tests validate proper fact extraction, relationship creation, and error handling.
-
-3. **Standardized Test Utilities**: The `test-utils.ts` file provides a common foundation for all tests with robust mock factories and type definitions.
-
-4. **Event Handling Coverage**: The logic tests cover all major event types with proper error handling and edge cases.
-
-5. **Type Safety**: Most tests maintain strong typing with proper mocking of required interfaces.
-
-## Areas for Improvement
-
-1. **Mock Runtime Gaps**: The most significant issue is incomplete mock runtime implementations. While recent changes added missing methods like `getMemoriesByRoomIds` and `searchMemories`, there are still issues with methods not properly implementing the behavior expected by the plugin.
-
-2. **Plugin Initialization Tests**: The tests for the plugin initialization need repair. The mock initialization function doesn't properly invoke the actual methods.
-
-3. **Service Mocks**: Many service tests fail due to missing or incomplete service initialization.
-
-4. **Provider Implementation Tests**: The provider tests have several failures related to precise mock requirements and assertion mismatches.
-
-5. **Error Case Testing**: While error handling exists, some tests expect specific error responses that don't match the actual implementation.
-
-## Recommended Next Steps
-
-1. **Fix Plugin Test**: Repair the mock implementation in `plugin.test.ts` to properly call the actual registration methods.
-
-2. **Address Remaining Provider Issues**: Complete the mocking for provider tests, particularly for `choice`, `facts`, and `settings` providers.
-
-3. **Enhance Service Tests**: Improve the service test initialization and mocking to support proper service testing.
-
-4. **Standardize Error Responses**: Ensure error handling in tests matches the expected output from the actual implementation.
-
-5. **Integration Testing**: Consider adding higher-level integration tests that use fewer mocks for more realistic testing scenarios.
-
-## Usage
-
-Run all tests:
-
-```
-npx vitest run packages/plugin-bootstrap/__tests__
-```
-
-Run specific test file:
-
-```
-npx vitest run packages/plugin-bootstrap/__tests__/actions.test.ts
-```
-
-Run tests in watch mode:
-
-```
-npx vitest watch packages/plugin-bootstrap/__tests__
-```
-
-## Test Utilities
-
-The test suite includes a robust set of utilities in `test-utils.ts` that provide:
-
-- `createMockRuntime()`: Creates a comprehensive mock of the `IAgentRuntime` interface
-- `createMockMemory()`: Creates mock Memory objects for testing
-- `createMockState()`: Creates mock State objects for testing
-- `createMockService()`: Creates mock Service objects for testing
-
-These utilities should be preferred over creating custom mocks to ensure consistency across the test suite.
